+++ include/snet/cli/command.hpp
