+++ src/tls/CMakeLists.txt
@@ -1,0 +1,5 @@
+# Sources
+file(GLOB_RECURSE SOURCES *.cpp)
+
+# Set target library
+add_library(snet_tls STATIC ${SOURCES})
